Extract list commands to mixin
