+++ fakeredis/commands_mixins/generic_mixin.py
