+++ fakeredis/commands_mixins/scripting_mixin.py
